+++ src/_adapters.rs
@@ -1,0 +1,2 @@
+pub mod _mdbook;
+pub mod default;